⚡ Improved data handling for sse
