+++ services/stories.js
@@ -33,26 +33,6 @@ export const fetchThisStory = async (data) => {
   }
 }
 
-export const isStoryLiked = async (story, user) => {
-  try {
-    const response = await axios.get(`/api/story/like?storyId=${story}&userId=${user}`,
-      {
-        headers: {
-          'Content-Type' : 'application/json',
-        }
-      }
-    )
-
-    return {
-      status: 200,
-      data: response.data,
-    }
-  } catch (error) {
-    console.log(error);
-    throw new Error(error.response?.data?.message || "Failed to fetch 'isStoryLiked'")
-  }
-}
-
 export const likeStory = async (storyId, userId) => {
   try {
     const response = await axios.post('/api/story/like',
